requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
Gunicorn==20.1.0
Pillow==9.1.0
